+++ src/ts/EventHandlers/VPAIDEventHandler.ts
@@ -5,6 +5,7 @@ import { VPAIDEndScreen } from 'Views/VPAIDEndScreen';
 import { AbstractOverlay } from 'Views/AbstractOverlay';
 import { OperativeEventManager } from 'Managers/OperativeEventManager';
 import { ThirdPartyEventManager } from 'Managers/ThirdPartyEventManager';
+import { ComScoreTrackingService } from 'Utilities/ComScoreTrackingService';
 import { VPAIDCampaign } from 'Models/VPAID/VPAIDCampaign';
 import { Diagnostics } from 'Utilities/Diagnostics';
 import { DiagnosticError } from 'Errors/DiagnosticError';
@@ -15,6 +16,7 @@ export class VPAIDEventHandler implements IVPAIDHandler {
     private _nativeBridge: NativeBridge;
     private _operativeEventManager: OperativeEventManager;
     private _thirdPartyEventManager: ThirdPartyEventManager;
+    private _comScoreTrackingService: ComScoreTrackingService;
     private _adUnit: VPAIDAdUnit;
     private _vpaidEventHandlers: { [eventName: string]: () => void; } = {};
     private _vpaidCampaign: VPAIDCampaign;
@@ -24,17 +26,20 @@ export class VPAIDEventHandler implements IVPAIDHandler {
     private _overlay: AbstractOverlay;
     private _adDuration: number = -2;
     private _adRemainingTime: number = -2;
+    private _abGroup: number;
 
     constructor(nativeBridge: NativeBridge, adUnit: VPAIDAdUnit, parameters: IVPAIDAdUnitParameters) {
         this._nativeBridge = nativeBridge;
         this._operativeEventManager = parameters.operativeEventManager;
         this._thirdPartyEventManager = parameters.thirdPartyEventManager;
+        this._comScoreTrackingService = parameters.comScoreTrackingService;
         this._adUnit = adUnit;
         this._vpaidCampaign = parameters.campaign;
         this._placement = parameters.placement;
         this._vpaidView = parameters.vpaid;
         this._overlay = parameters.overlay;
         this._vpaidEndScreen = parameters.endScreen;
+        this._abGroup = parameters.configuration.getAbGroup();
 
         this._vpaidEventHandlers.AdError = this.onAdError;
         this._vpaidEventHandlers.AdLoaded = this.onAdLoaded;
@@ -101,6 +106,14 @@ export class VPAIDEventHandler implements IVPAIDHandler {
         }
     }
 
+    public getDuration() {
+        return this._adDuration;
+    }
+
+    public getPlayTime() {
+        return (this._adDuration - this._adRemainingTime) * 1000;
+    }
+
     private onAdLoaded() {
         this._adUnit.getAdUnitNotLoadedTimer().stop();
         this.onVPAIDProgress(this._adDuration, this._adRemainingTime);
@@ -132,6 +145,9 @@ export class VPAIDEventHandler implements IVPAIDHandler {
         this._nativeBridge.Listener.sendStartEvent(this._placement.getId());
         this._adUnit.sendTrackingEvent('creativeView');
         this._operativeEventManager.sendStart(this._adUnit);
+        if (this._abGroup === 5) {
+            this.sendComscoreEvent('play', 0);
+        }
     }
 
     private onAdImpression() {
@@ -162,6 +178,9 @@ export class VPAIDEventHandler implements IVPAIDHandler {
         this._adUnit.sendTrackingEvent('complete');
         this._adUnit.setFinishState(FinishState.COMPLETED);
         this._operativeEventManager.sendView(this._adUnit);
+        if (this._abGroup === 5) {
+            this.sendComscoreEvent('end', (this._adDuration - this._adRemainingTime) * 1000);
+        }
     }
 
     private onAdPaused() {
@@ -204,5 +223,19 @@ export class VPAIDEventHandler implements IVPAIDHandler {
         for (const url of urls) {
             this._adUnit.sendThirdPartyEvent('vpaid video click', url);
         }
+    }
+
+    private sendComscoreEvent(eventName: string, position: number) {
+        const sessionId = this._adUnit.getCampaign().getSession().getId();
+        const creativeId = this._adUnit.getCampaign().getCreativeId();
+        const category = this._adUnit.getCampaign().getCategory();
+        const subCategory = this._adUnit.getCampaign().getSubCategory();
+        let adDuration = (this._adDuration * 1000);
+
+        if (adDuration < 0 || typeof adDuration === 'undefined') {
+            adDuration = 0;
+        }
+
+        this._comScoreTrackingService.sendEvent(eventName, sessionId, adDuration.toString(10), position, creativeId, category, subCategory);
     }
 }